refactor: merge pty_linux.go and pty_darwin.go into pty_unix.go
Both files became identical after enabling credential demotion on macOS
and adding the non-root guard. Merged into pty_unix.go with
//go:build !windows tag for future Windows support.

Also skip credential demotion when running as non-root, matching the
pattern used by utils.Demote(). This allows PTY sessions to work in
development mode without root privileges.
